Adding basic JavaScript DOM manipulation codes
Here the app.ts adds a class which initializes the form of the HTML.
The class - ProjectInput - has three attributes which are HTMLElements,
but are more specific than the 'HTMLElements' themselves. Specifically,
the three attributes are template, div, and form elements, which are
found in the index.html file.

The constructor imports these elements from index.html and assigned them
to the attribute. The attributes are type casted in the end by using the
syntax 'as HTMLDivElement' as an example. However, this can also be set
as <HTMLDivElement>document.getElementById('app'). The exclamation mark
at the end tells TypeScript that we are sure there will be this element
because without it, it can be null, or the div element itself. As a side
note, it's possibel that this could be assiged with ?. (optional
chaining)

Both document.importNode, firstElementChild, and insertAdjacentElement
are JavaScipt DOM functions, which needs to be reviewed since this is
currently unfamiliar.