Adding addEventListener to title Input
Three more input elements were added to the class attribute: title,
description, and people, which are all inside the template element.

The constructor, again, grabs the elements from HTML, and two new
functions were added to the class so that it can modify them. The
first function, submitHandler, is basically console.log the value of the
element. The second function, configure, adds the event listener to the
element submit is pressed. One thing to note that the bind function was
attached since submitHandler would not know what 'this' is. It will need
to be attached to 'element' attribute of the class, and hence the bind
function was added

Later on, instead of bind, this will be converted using decorators