reformating validation function
All of the inputs - title, description, and people - have now been
validated using a function instead of being checked on the form.

Interface is first created for the input field, which includes mix of
value types (either strings from title and description, and number from
people). Each "if" statements are checking the field of interface, and
also depends on whether they are included or not, since most of the
fields are optional.

The minLength and maxLength checks the string length, while min and max
checkes the number of people.

Afterwards, the gatherUserInput checks the actual inputs from index.html
and then calls the validate function to verify that they meet the
requirements. Note that the value of min and max are set in index.html
where input parameter inside the div id=form-control. These values are
set manually here.