adding project list render class
Initial setup for rendering project list, which are separated into two
parts: active and finished. Both lists are rendered which means
initially there are empty lists for both sections.

The setup is similar to the project input class, but the DOM query is
on the render list inside the index.html, which is under the template
with id='project-list'.

There still needs clarification of insertAdjacentElement function
but this should be reviewed at the end of the project.